Support MMYY MM.YY MM-YY short formats
The MMYY format also supports MYY and was directly requested. The MM.YY and
MM-YY formats were trivial to add support for, hence why they are also added.

As part of this commit I have also moved the DATE_PATTERN as a constant to the
date_parse.js file since that information is more pertinent in that location.

Note: The MMYY format does not support MYY as entering `101` would immediately
convert into `Jan 2001` preventing `1010` for `Oct 2010` from being entered.